[FLINK-13502] [table-planner-blink] move CatalogTableStatisticsConverter to planner.utils
